+++ pkg/agent/storage/ci/keyvalue_store_ci.go
@@ -1,38 +1,19 @@
-package storage
+package storageci
 
 import (
 	"os"
-	"path/filepath"
 	"testing"
 
 	"github.com/go-kit/kit/log"
+	agentbbolt "github.com/kolide/launcher/pkg/agent/storage/bbolt"
+	"github.com/kolide/launcher/pkg/agent/storage/inmemory"
 	"github.com/kolide/launcher/pkg/agent/types"
-	"github.com/stretchr/testify/require"
-	"go.etcd.io/bbolt"
 )
 
-const (
-	dbTestFileName = "test.db"
-)
-
-func NewCIKeyValueStore(t *testing.T, logger log.Logger, bucketName string) (types.KVStore, error) {
+func NewStore(t *testing.T, logger log.Logger, bucketName string) (types.KVStore, error) {
 	if os.Getenv("CI") == "true" {
-		return NewInMemoryKeyValueStore(logger), nil
+		return inmemory.NewStore(logger), nil
 	}
 
-	return NewBBoltKeyValueStore(logger, setupDB(t), bucketName)
-}
-
-func setupDB(t *testing.T) *bbolt.DB {
-	// Create a temp directory to hold our bbolt db
-	dbDir := t.TempDir()
-
-	// Create database; ensure we clean it up after the test
-	db, err := bbolt.Open(filepath.Join(dbDir, dbTestFileName), 0600, nil)
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		require.NoError(t, db.Close())
-	})
-
-	return db
+	return agentbbolt.NewStore(logger, agentbbolt.SetupDB(t), bucketName)
 }
